requirements.txt
# conda create -n myenv python=3.10
# conda activate myenv
# pip install -r requirements.txt

pandas==1.5.3
numpy==1.23.5
python-Levenshtein==0.27.1
fuzzywuzzy==0.18.0
openpyxl==3.1.2
python-docx==1.1.2